Fix Bug [#2089] maxObjects stored as int in table configurations: should be stored as bigint

+++ harvester/harvester/src/main/java/dk/netarkivet/harvester/datamodel/DomainDBDAO.java
@@ -489,7 +489,7 @@ public class DomainDBDAO extends DomainDAO {
                     // Update
                     DBUtils.setComments(s, 1, dc, Constants.MAX_COMMENT_SIZE);
                     s.setString(2, dc.getOrderXmlName());
-                    s.setInt(3, dc.getMaxObjects());
+                    s.setLong(3, dc.getMaxObjects());
                     s.setInt(4, dc.getMaxRequestRate());
                     s.setLong(5, dc.getMaxBytes());
                     s.setString(6, dc.getName());
@@ -813,7 +813,7 @@ public class DomainDBDAO extends DomainDAO {
             DBUtils.setComments(s, 2, dc, Constants.MAX_COMMENT_SIZE);
             s.setLong(3, d.getID());
             s.setLong(4, templateId);
-            s.setInt(5, dc.getMaxObjects());
+            s.setLong(5, dc.getMaxObjects());
             s.setInt(6, dc.getMaxRequestRate());
             s.setLong(7, dc.getMaxBytes());
             int rows = s.executeUpdate();
